+++ intellij-plugin-structure/structure-intellij/src/main/java/com/jetbrains/plugin/structure/intellij/plugin/loaders/JarPluginLoader.kt
@@ -9,10 +9,13 @@ import com.jetbrains.plugin.structure.base.problems.UnableToExtractZip
 import com.jetbrains.plugin.structure.base.problems.UnableToReadDescriptor
 import com.jetbrains.plugin.structure.base.utils.getShortExceptionMessage
 import com.jetbrains.plugin.structure.base.utils.simpleName
+import com.jetbrains.plugin.structure.base.zip.ZipArchiveException
+import com.jetbrains.plugin.structure.base.zip.newZipHandler
 import com.jetbrains.plugin.structure.intellij.plugin.IdePluginManager.Companion.META_INF
 import com.jetbrains.plugin.structure.intellij.plugin.PluginCreator
 import com.jetbrains.plugin.structure.intellij.plugin.PluginCreator.Companion.createInvalidPlugin
 import com.jetbrains.plugin.structure.intellij.plugin.PluginCreator.Companion.createPlugin
+import com.jetbrains.plugin.structure.intellij.problems.DuplicateEntryInJar
 import com.jetbrains.plugin.structure.intellij.problems.PluginCreationResultResolver
 import com.jetbrains.plugin.structure.intellij.resources.ResourceResolver
 import com.jetbrains.plugin.structure.jar.JarArchiveCannotBeOpenException
@@ -29,6 +32,16 @@ private val LOG: Logger = LoggerFactory.getLogger(JarPluginLoader::class.java)
 
 internal class JarPluginLoader(private val fileSystemProvider: JarFileSystemProvider) : PluginLoader<JarPluginLoader.Context> {
   override fun loadPlugin(pluginLoadingContext: Context): PluginCreator = with(pluginLoadingContext) {
+    try {
+      val duplicates = jarPath.newZipHandler().findDuplicateEntries()
+      if (duplicates.isNotEmpty()) {
+        val firstDuplicate = duplicates.first()
+        LOG.warn("Duplicate ZIP entry '{}' found in [{}]", firstDuplicate, jarPath)
+        return createInvalidPlugin(jarPath, descriptorPath, DuplicateEntryInJar(firstDuplicate, jarPath.simpleName))
+      }
+    } catch (e: ZipArchiveException) {
+      LOG.warn("Unable to scan [{}] for duplicate entries: {}", jarPath, e.getShortExceptionMessage())
+    }
     return try {
       PluginJar(jarPath, fileSystemProvider).use { jar ->
         when (val descriptor = jar.getPluginDescriptor("$META_INF/$descriptorPath")) {
@@ -50,6 +63,7 @@ internal class JarPluginLoader(private val fileSystemProvider: JarFileSystemProv
               createInvalidPlugin(jarPath, descriptorPath, UnableToReadDescriptor(descriptorPath, message))
             }
           }
+
           PluginDescriptorResult.NotFound -> createInvalidPlugin(
             jarPath,
             descriptorPath,