+++ intellij-plugin-structure/structure-intellij/src/main/java/com/jetbrains/plugin/structure/intellij/problems/PluginCreationResultResolver.kt
@@ -81,6 +81,7 @@ class IntelliJPluginCreationResultResolver : PluginCreationResultResolver {
     OptionalDependencyConfigFileIsEmpty::class,
     DependencyConstraintsDuplicates::class,
 
+    DuplicateEntryInJar::class,
     PluginLibDirectoryIsEmpty::class,
     PluginZipContainsMultipleFiles::class,
     PluginZipContainsSingleJarInRoot::class,